댓글 리스트 양식. 댓글 작성 처리 기능 구현중

+++ src/main/java/com/kjh/exam/demo/controller/UsrArticleController.java
@@ -55,7 +55,7 @@ public class UsrArticleController {
 
 		ResultData<Integer> writeArticleRd = articleService.writeArticle(rq.getLoginedMemberId(), boardId,  title, body);
 
-		int id = (int) writeArticleRd.getData1();
+		int id = writeArticleRd.getData1();
 		
 		Article article = articleService.getArticle((int) writeArticleRd.getData1());
